+++ src/sim/simulation/actions/move/go_to_deposit.py
@@ -2,19 +2,20 @@ import logging
 
 from simpy import Environment
 
+from src.sim.drawer import Drawer
+from src.sim.simulation.actions.move.move import Move
+from src.sim.simulation.simulation import Simulation
 from src.sim.status_warehouse.enum_warehouse import EnumWarehouse
-from src.sim.simulation import Simulation
-from src.sim.status_warehouse.simulate_events.move.move import Move
-from src.sim.warehouse import Warehouse, Drawer
+from src.sim.warehouse import Warehouse
 
 logger = logging.getLogger(__name__)
 
 
-class Unload(Move):
+class GoToDeposit(Move):
     def __init__(self, env: Environment, warehouse: Warehouse, simulation: Simulation, drawer: Drawer,
                  destination: EnumWarehouse):
         """
-        Horizontal movement from the center of the column to the center of the lift.
+        Movement to go to the deposit.
 
         :type env: Environment
         :type warehouse: Warehouse
@@ -29,9 +30,7 @@ class Unload(Move):
         """
         super().__init__(env, warehouse, simulation, destination, drawer)
 
+    # override
     def simulate_action(self):
-        logger.debug(f"Time {self.env.now:5.2f} - Start unloading a drawer")
-        yield self.env.process(self.simulation.unload(
-            self.get_drawer(),
-            rmv_from_cols=True if self.get_destination() == EnumWarehouse.CAROUSEL else False)
-        )
+        logger.debug(f"Time {self.env.now:5.2f} - Start come back to deposit position")
+        yield self.env.process(self.simulation.go_to_deposit())