+++ src/sim/simulation/actions/move/load.py
@@ -2,19 +2,19 @@ import logging
 
 from simpy import Environment
 
-from src.sim.simulation import Simulation
+from src.sim.simulation.actions.move.move import Move
+from src.sim.simulation.simulation import Simulation
 from src.sim.status_warehouse.enum_warehouse import EnumWarehouse
-from src.sim.status_warehouse.simulate_events.move.move import Move
 from src.sim.warehouse import Warehouse, Drawer
 
 logger = logging.getLogger(__name__)
 
 
-class Vertical(Move):
+class Load(Move):
     def __init__(self, env: Environment, warehouse: Warehouse, simulation: Simulation, drawer: Drawer,
                  destination: EnumWarehouse):
         """
-        Vertical movement.
+        Horizontal movement from the center of the lift to the center of the column.
 
         :type env: Environment
         :type warehouse: Warehouse
@@ -30,8 +30,5 @@ class Vertical(Move):
         super().__init__(env, warehouse, simulation, destination, drawer)
 
     def simulate_action(self):
-        logger.debug(f"Time {self.env.now:5.2f} - Start vertical move")
-        if self.get_destination() == EnumWarehouse.CAROUSEL:
-            yield self.env.process(self.simulation.reach_drawer_height(self.get_drawer()))
-        else:
-            yield self.env.process(self.simulation.allocate_best_pos(self.get_drawer()))
+        logger.debug(f"Time {self.env.now:5.2f} - Start loading inside the warehouse")
+        yield self.env.process(self.simulation.load(self.get_drawer(), self.get_destination()))