+++ src/sim/simulation/actions/move/unload.py
@@ -2,19 +2,19 @@ import logging
 
 from simpy import Environment
 
+from src.sim.simulation.actions.move.move import Move
+from src.sim.simulation.simulation import Simulation
 from src.sim.status_warehouse.enum_warehouse import EnumWarehouse
-from src.sim.simulation import Simulation
-from src.sim.status_warehouse.simulate_events.move.move import Move
 from src.sim.warehouse import Warehouse, Drawer
 
 logger = logging.getLogger(__name__)
 
 
-class Load(Move):
+class Unload(Move):
     def __init__(self, env: Environment, warehouse: Warehouse, simulation: Simulation, drawer: Drawer,
                  destination: EnumWarehouse):
         """
-        Horizontal movement from the center of the lift to the center of the column.
+        Horizontal movement from the center of the column to the center of the lift.
 
         :type env: Environment
         :type warehouse: Warehouse
@@ -30,5 +30,8 @@ class Load(Move):
         super().__init__(env, warehouse, simulation, destination, drawer)
 
     def simulate_action(self):
-        logger.debug(f"Time {self.env.now:5.2f} - Start loading inside the warehouse")
-        yield self.env.process(self.simulation.load(self.get_drawer(), self.get_destination()))
+        logger.debug(f"Time {self.env.now:5.2f} - Start unloading a drawer")
+        yield self.env.process(self.simulation.unload(
+            self.get_drawer(),
+            rmv_from_cols=True if self.get_destination() == EnumWarehouse.CAROUSEL else False)
+        )