+++ src/sim/simulation/actions/move/vertical.py
@@ -2,20 +2,19 @@ import logging
 
 from simpy import Environment
 
+from src.sim.simulation.actions.move.move import Move
+from src.sim.simulation.simulation import Simulation
 from src.sim.status_warehouse.enum_warehouse import EnumWarehouse
-from src.sim.drawer import Drawer
-from src.sim.simulation import Simulation
-from src.sim.status_warehouse.simulate_events.move.move import Move
-from src.sim.warehouse import Warehouse
+from src.sim.warehouse import Warehouse, Drawer
 
 logger = logging.getLogger(__name__)
 
 
-class GoToDeposit(Move):
+class Vertical(Move):
     def __init__(self, env: Environment, warehouse: Warehouse, simulation: Simulation, drawer: Drawer,
                  destination: EnumWarehouse):
         """
-        Movement to go to the deposit.
+        Vertical movement.
 
         :type env: Environment
         :type warehouse: Warehouse
@@ -30,7 +29,9 @@ class GoToDeposit(Move):
         """
         super().__init__(env, warehouse, simulation, destination, drawer)
 
-    # override
     def simulate_action(self):
-        logger.debug(f"Time {self.env.now:5.2f} - Start come back to deposit position")
-        yield self.env.process(self.simulation.go_to_deposit())
+        logger.debug(f"Time {self.env.now:5.2f} - Start vertical move")
+        if self.get_destination() == EnumWarehouse.CAROUSEL:
+            yield self.env.process(self.simulation.reach_drawer_height(self.get_drawer()))
+        else:
+            yield self.env.process(self.simulation.allocate_best_pos(self.get_drawer()))