+++ src/web_app/layouts/simulation_statistics.py
@@ -4,8 +4,8 @@ import dash_bootstrap_components as dbc
 import plotly.graph_objs as go
 from dash import html, dcc
 from pandas import DataFrame
+from src.sim.status_warehouse.actions.action_enum import ActionEnum
 
-from src.sim.status_warehouse.simulate_events.action_enum import ActionEnum
 from src.sim.utils.statistics.warehouse_statistics import WarehouseStatistics, TimeEnum