+++ tests/sim/utils/statistics/test_warehouse_statistics.py
@@ -3,8 +3,8 @@ from unittest import TestCase
 from numpy import int64
 from pandas import DataFrame, Timestamp, Timedelta
 
-from src.sim.simulation import Simulation
-from src.sim.status_warehouse.simulate_events.action_enum import ActionEnum
+from src.sim.simulation.actions.action_enum import ActionEnum
+from src.sim.simulation.simulation_type.warehouse_simulation.warehouse_simulation import WarehouseSimulation
 from src.sim.utils.statistics.warehouse_statistics import WarehouseStatistics, TimeEnum
 from src.sim.warehouse import Warehouse
 
@@ -12,7 +12,7 @@ from src.sim.warehouse import Warehouse
 class TestWarehouseStatistics(TestCase):
     def setUp(self):
         self.warehouse = Warehouse()
-        self.simulation = Simulation(self.warehouse)
+        self.simulation = WarehouseSimulation(self.warehouse)
         self.simulation.run_simulation()
         self.warehouse_statistics = WarehouseStatistics(DataFrame(self.simulation.get_store_history().items))